Remove mdspell/markdown-spellcheck
**Why**:

- The pre-commit hook we are using appears to have never worked since
it was introduced in November 2019 because it is supposed to ignore
numbers and acronyms, yet our `.spelling` file has several instances of
those.
- Our `.spelling` file is over 600 lines long, which shows you how many
false positives we've had to manually ignore just to be able to commit
an ADR! Think of all the time wasted doing this.
- A misspelling should not prevent us from making a commit message.
What's the worst that can happen? Let's trust each other to find it
during PR review, or to fix it later while reading an ADR.